Replaced all java common logging with slf4j (#6)
Finally did some reading on how to do Java logging well*, I changed most of the interfaces to SLF4J (simple logging facade 4 java, names are horrible), which is just a standard interface on top of many different logging backends.

We still use Java common logging (JCL) under the hood, so nothing's changed there yet. But we have the flexibility to when we want to, say, if we need to start running multiple proxys and a load balancer to them, some distributed logging system would work much better in that case. But for now, this is good. 

* I liked https://sematext.com/blog/java-logging/ a lot