Added all the MDE Ports (except court scheduling)
By running wsdl2java on all of the SOAP port MDEs, not
just FilingReview.

NOTE(brycew): by all, we mean everything but the "CourtScheduling" MDE,
which has items from the ECF v5, which will override some existing
classes, likely making them not compatiable with the ECF v4 SOAP stuff
(which is pretty strict). If this feature is needed, we'll have to put
it in another namespace.

This also adds the important `binding.xjb` officially to the repo.

This is a work in progress, but it fixes the annoying issue that looks
like:

> Simple type "VMOCodeSimpleType" was not mapped to Enum due to
> EnumMemberSizeCap limit. Facets count: 1,121, current limit: 1,024.
> You can use customization attribute "typesafeEnumMaxMembers" to
> extend the limit

However, I can't find a way to get the bindings.xjb to apply to more
than one file at a time (you get another error about "compiler was
unable to honor this globalBindings customization.") So for now
it's applying to the file, `ecf-11.xsd`, which is the one with
languages, and not anything else.