Guess what: more ecf5 / NIEM code!
The GetReturnDate docs describe a civil:CaseAugmentation XML element
that is necessary, but wasn't anywhere. Turns out, it's in the civil.xsd on the
server, that is referenced absolutely nowhere in the rest of the ECF v5 schema.
So I downloaded it and the Human Resources Schema separately and added them to
the existing schema files. They don't follow the same numerical naming, since
they are manual additions.