Fix bug in LoginDatabase and new Payment Accounts
LoginDatabase would just return 200 if you had a correct
API key, but didn't actually log in to any efilers. So
that has been changed to return a 403 if nothing was provided.

The SOAP objects for Payment Accounts have been changed,
so they need to also be changed in the Service.java file.