HATEOS (#96)
* An initial go at adding navigable endpoints for the server.

Can go to the root of the site, get a list of services,
go to those services, and then see what you can do with them.

Ideally, it would work like `https://api.github.com`. We aren't quite
there yet, because I don't want to solidify the attributes yet, but
we're getting closer. Biggest thing in the way there are GET query params
that shouldn't exist and end points on individual objects, like courts.

* comments
* Got all levels of paths in codes service working

Bigger part: mini tests for the services are in and working now