Updated payments to work with Tyler's new API
Description of New functionality is in EFM-3616 in the release notes of
Feb 25, 2022. Essentially, they need a CodeId, not the text-based code anymore.

Not documented is the fact that it needs to be both in the `PaymentAccountTypeCode`
attribute (even though it's a string in the XML schema) and the
`PaymentAccountTypeCodeId` in the body. Another common screw up is that Java
will let you directly add the JAXElement<Integer> from the GetPaymentAccountTypeList
directly to the CreatePaymentAccount. DON'T do this, as it has the wrong QNAME,
and you'll get an Error 44: "AccountTypeId '' did not match any account type."