Handle already-in-case parties in the filing
Necessary because DA will require any party that you reference in a generated
PDF to be in the `users` or `other_parties` struct. We will refer to those users
from the service contacts, filing parties, and attorney assignments
if they are there, even if they aren't new.

Required a descent amount of refactoring, and removed a lot of information
from the `checking_filtered_users.json` that wasn't actually needed.