Email Templates and Document Attachments
A bigger PR that combines two different features:

This PR adds the ability to provide email templates in the filing submission itself,
set (ideally) in Docassemble. This is easier for most DA authors than using
the raw REST API to change the email templates, which doesn't have a good user
interface on yet.

There are 4 different email templates that can be set:
* `email_confirmation_contents`, which is the email that is send immediately after
  the filing is submitted
* `acceptance_contents`: sent when the clerk acts on the submitted filing, only
  if the action is `accept`
* `rejected_contents`: sent when the clerk acts on the submitted filing, only
  if the action is `rejected`
* `neutral_contents`: sent when the clerk acts on the submitted filing with any
  other action. This includes a lot of things that might be postiive or negative
  to the filer, like "under-review", "moot", "granted", "failed", "denied", and
  "error".

The email templates are saved in the database when the filing is submitted, so
the email template is fixed at time of submission, and can't be updated after
that.

This PR also adds the ability to split a single document into several
"attachments" (really just individual PDFS). This is useful for forms that
really have different components, like a fee waiver that has the application and
an additional court order. Tyler has always had the ability to support this, but
our implementation was simplified.

Now, the proxy server will go two levels deep in the `al_court_bundle`. The
first level are the filings themselves, like `Appearance`, and `Fee Waiver`. The
next level are the documents that make up those filings, like `Fee Waiver
application` and `Fee Waiver Order`.

The 2nd level documents need to have the `filing_component` (lead or attachment)
and `document_type` (usually public or confidential) attributes set. An error
will occur later in the process if these documents are present but not actually
supposed to be used. If you don't want to use the 2nd level documents when
filing, and only want a single PDF to be submitted (if, for example, you are
filing in the "St. Claire" court, which doesn't allow the `Attachment`
filing_component, effectively forcing you to only submit one document), then you
should set the `tyler_merge_attachments` attribute to true on the 1st level
document.