Added ACME challenge library and classes
Can start the challenge from a separate program running, and then detect
and publish the token from the main server. Flexible enough that it should be
able to run on a timer automatically when you don't have to interactively agree
to the terms of service.

Uses acme4j, with most of the AcmeRenewal class coming from
https://shredzone.org/maven/acme4j/example.html.