Spotbug (#126)
Spotbug is a java linter. IMO, we need a java linter, since:
* as a team, we don't have a ton of java knowledge about best practices
* I know a bit of java best practices, but am not used to handling repos of this size solo
* anything we can do to reduce anxiety about existing bugs in the code and making new releases, I want to do
* I will one day leave

I didn't fix all of the found spotbugs because:
* there are quite a few
* many are in autogenerated code
* some would require pretty substantial changes
* I just don't agree with some. I use capitalized static methods as constructors sometimes (`MyClass.SpecialConstructor`), so I can return `Optional.empty` if need be, and now have throw exceptions. Capitalizing those makes more sense to me, as they are acting as constructors, and not as methods.

This is fine; I'll keep using this as need be, and won't integrate into the CI just yet. It would have to be something more than a "pass/fail" workflow, and would need to be closer to coverage (you're this much better!), which is a bit more complicated than what I've done so far, and would need to research it.

To run spotbug, just use `mvn spotbugs:spotbugs`. After that finishes, you can view the linter errors and warnings in a gui with `mvn spotbugs:gui`.