All ECF5 work
Working: all calls, including filing through GetFees and FilingReview.

Noting that Charge Indicator needs to be true for things to work (that or
tyler case aug needs to come before civil case aug).

Passing tests now too: Partially because of an internal CD refactor that broke mocking, and
partially because of Endpoint reflection not considering abstract classes
and JAX annotatinos on them (annonations can't be on subclasses, or they have
to be repeated, defeating the point of the parent class).

Removed optional services as well?