Clean code updater (#133)
Had to use this class a lot recently when trying to debug things like refundreason not actually downloading. The issue was that as a thing to download, it's just not in the Policy at all, it's only showing up in the versions list. We can't
assume they'll be the same anymore, and I had to code around a bit of that.