Smaller codes DB tables (#138)
Shrinks the code tables, particularly the optional services one, which required some refactoring to not assume that most operations take only a single transaction.

It shrinks the disk usage by a few GB, can't find my notes that I have for how much. It wasn't a ton though, and would need to go to enums if we wanted to actually decrease usage.

For now though, addresses #99.

* Update 6to7, smaller database than otherwise

More progress; still fails test from nothing

Had to pivot again, and start putting a lot of specific DELETE, INSERT,
and CREATE operations into the optional service class, since
that table is now two tables, which means two statements, which breaks all
of the abstractions we had that adding to a single codes table is adding to
a single SQL table. Just special cased a lot of stuff for Optionalservices.

Passed all tests, still need to local test

* Removed optionalservices from table list, not used anymore
* DataFieldRow neews booleans now, not strings
* Fix tests for optsrv, add test for deleting
* Return after deleting from optionalservices

Instead of continuing to non-optional service stuff.