Dedupe optionalservice table
Despite the fact that there are two tables for optional services to shrink
the disk space usage, there was an error that would duplicate entries in the
table that should have been much smaller, but it would only happen on insert/
a version update for the court. No logical errors happened, but would result in
a large amount of codes and extra parsing to be done on the docassemble side.