Database refactor: make all AutoCloseable
Makes the interface for the various database classes that we use AutoCloseable.
This results in a lot of small changes everywhere, as we don't have to access
the underlying Connection object for convience functions or to close it.

As a result of some of these changes, we also start giving Services a
`Supplier<CodeDatabase>`. This also preps for ECF5; the CodesDatabase will have
two implementations, one for ECF4 and one for ECF5, but the services won't
know which to use without lots of extra configs or globals. So instead we
simply give them a Supplier, that they can call and use as normal; the
constructing object knows which of ECF4 vs ECF5 to use.