Changed CodesService class hierarchy
Added more features to the CodeService interface, such as a CORS function, to
let users get the codes from JS in the browser. For this to work, I also needed
to update Endpoint Reflection to work on Interfaces and parent classes as well
as just the current class (JAX-RS already works with parent classes).

Some side effects:
* started passing in a "passwordChecker" lambda to AdminUserService.
* removed a lot of annotations from certain services
* Cleaned up CodeDatabases, and added a LanguageNames function
* reorganized Map entries in CodeUpdater
* some other small clean ups in Ecf4Filer, etc.