Show current values of wrong vars for errors
In order to effectively debug errors that happen, we need to know the current state
of the interview (it's easier than trying to evaluate all the ways a docassemble
interview could get the wrong value). This adds a `currentVal` attribute to
`InteriewVariable`, which is returned whenever there is an error in the provided
filing. Defaults to `Optional.empty()`.

Also added unit tests with jqwik to check that we can serialize InfoCollection to
JSON reliably, and properly escaped JSON strings for that reason.