Removed redundant call to wrong FirmService (#236)
`ServiceHelpers` had an extra call to an old API that wasn't correctly
removed in #217, and `TylerFirmClient used the wrong class attribute
when calling a version of the `2022_1` API. Wasn't caught because it
only runs on (and only be run on) Massachusetts PROD, and PROD isn't a
normal part of our test suite [^1].

Found after monitoring a push to production today, and was able to
recreate and fix locally. Some parts of the `FirmService` wrapper that
weren't working. Checked the rest of the repo for similar errors and
didn't find any.

[^1]: This should be fixed soon, when we introduce the mock server stuff
(should mock out prod servers at older versions of the API).