Better Codes API docs
Added all codes endpoints to the abstract `CodesService` class so enunciate could pick
them up for documentation. They are marked as `Unimplemented` in the JeffNet version (which is deprecated and will
be removed soon anyway).

All of the implementations in `EcfCodesService` have been marked with `@Override` and had the JAX-RS annotations
removed, but not functionality has changed. Confirmed that enunciate can pick up javadocs from classes earlier
in the path. The `@responseExample` javadoc annotation isn't great, but gets the job done.

Will get this merged, and will then work on documenting the rest of the parameters as needed.