Handle JSON parsing better
Found this by passing an invalid `RegistrationRequestType` to the `registerUser` endpoint (it needs to be all caps).

But for some reason, instead of erroring out, the server would return a 204, making it seem like some sort of user
had been registered. Was able to track the 204 down to the fact that CXF catches IOExceptions
(JAXRSInvoker.java:L287), which the JSON exception is considered a part of, but there were no exception handlers to
handle the exception, so it was treated as empty.

The fix involved making an exception handler to properly handle and category the exception as a user input
exception (400).

Also added a unit test on the AdminUserService as a regression test for this. Had to update mockito to
mock static methods, and refactored a small bit of the EfspServer setup into a static method so we can
easily use the same providers when testing as we do on the real server.