Stop passsing around DataSources, use Suppliers
i.e. Supplier<LoginDatabase>, Supplier<UserDatabase>, etc.

Previously, we were passing around the direct `DataSource` object, which is
an argument for the Database object (`LoginDatabase`, `UserDatabase`) constructors.
Works okay, but not easy to mock or handle what behaviors we really want from
the Database objects.

By using suppliers for the actual database objects, it's easier to mock,
and has a lower surface area for what methods a class can use on db anyway (so,
ideally less likely to make a mistake in future refactors).

Improve test mocking for AdminUserServiceTest