Make an actual User/Firm Factory
Separates out code that we wrote from the `EfmFirmService`/`EfmUserService`s,
into a specific factory class. Necessary for testing / mocking.

Since static mocking doesn't work across threads, in order to test services,
we need to be able to mock the TylerFirmClient/TylerUserClient to actually
mock Tyler's responses. Now that we do, we can test much of the `registerUser`
method's behavior.