Force passwords to be passed to register user
Because Tyler doesn't send back a useful error message, we need to be able to
individually identify what issues we have with an input. This does so for the password
attribute.

* Switches to a list of custom record types to iterate over because Map's error out when
the values are null, which we are trying to check to avoid. Record's won't error and can
keep the right return type.