Tyler Factories + Better missing password error response (#265)
Adds one refactor, fixes one bug.

* Refactor: Separates out code that we wrote from the
`EfmFirmService`/`EfmUserService`s, into a specific factory class.
Necessary for testing / mocking. Since static mocking doesn't work
across threads, in order to test services, we need to be able to mock
the TylerFirmClient/TylerUserClient to actually mock Tyler's responses.
Now that we do, we can test much of the `registerUser` method's
behavior.
* Bug: Because Tyler doesn't send back a useful error message, we need
to be able to individually identify what issues we have with an input.
This does so for the password attribute.
* Switches to a list of custom record types to iterate over because
Map's error out when the values are null, which we are trying to check
to avoid. Record's won't error and can keep the right return type.