Show more context with tyler errors
Some tyler errors (like "Lower court not found") aren't possible to
easily debug further without additional information. While some of this info 
might be able to be traces back to DA servers with request ids,
the most relevant information should be displayed in the log itself.

Currently only focuses on `Ecf4Filer.java`, but adds a method that takes
the `FilingInformation` already parsed from a filing, and logs the relevant
parts of that in regards to the Tyler errors codes.