Don't pass nulls instead of env
Wonderful choice by me only a few months ago to pass null instead of the TylerEnv
to the CodeDatabase when starting up, just to make tables; switching to an actual
object means we null pointer exeption immediately.

Not sure how no tests didn't catch this, will need to fix that.