Pass `CodeUpdater` a Supplier, not a CodeDatabase
Some operations (specifically `downloadIndividual` [^1]) don't
need the CodeDatabase. However, the SQL data source is always
created, so you need valid env vars for a SQL database, even
if not used.

This patch changes it so that isn't the case; if the operation
you're trying to run manually doesn't use a SQL database, the
variables are not required.

[^1]: `downloadInidivdual` gets individual XML files from
courts (necessary for `repcap`, an undocumented table which
breaks things), and doesn't directly interact with any
of the SQL tables themselves.