Adding court coverage to code search endpoints (#309)
Would allow a new param to let you either get all of the search results
for a string, or would let you get courts that a particular search query
would be sufficient for partial matches with.

Turns a custom python script that takes 150 seconds to run into a single
query that can be added to the codes interview to give detailed
information about a particular search string. That query will be added
once this is in stage.