Actually put Court Scheduling code in prod resources
Court Scheduling MDE WSDL files were never actually added for IL PROD,
probably because that feature wasn't used in production at all.

With the stricter start time checks enforced in the previous commit,
the prod server would error out since one of the services it's supposed
to load wouldn't be able to load.

This commit copies over all of the XSD files for the pre-release ECF-5
in IL needed for Court Scheduling MDE to the prod resources. It's a bit
hacky, but haven't had time to do the full refactor of how the XSDs are
handled yet (see #282 for more details).