Fix prod bug with spaces on the end of phone numbers  (#336)
A regex for the MA Appeals court was failing on a phone number because one
of the two possible numbers from a user had a space at the end that wasn't stripped.

Fixed in a few regards:

* adds the regex to the end of the error description (would have saved
me a bit of time when triaging)
* strips spaces off the end of the phone number if it doesn't pass the
regex the first time
* if at least one number that the user gives is valid, use it, and just
discard the others.

Related to the recent production alerts.