Ecfv5 Generated Files (#338)
Way too big to really review, but it's not executing any code yet. All
it's doing is:

* making it so that the Xsd Downloader can handle ECFv5 XML files
(different named files could have the same schema contents, can't have
duplicate schemas, so we have to hash the contents of the file when
downloading them)
* Add a bunch of stuff to the `redo_wsdl.py` file to closer match what
we're doing when downloading XML / generating Java
* not perfect yet, as there's still a lot of manual java namespace
re-naming that we have to do that isn't supported by `wsdl2java`.
Ideally, this will be baked into the python at some point (see #337)
* Make the shell / abstraction layers for handling multiple versions of
the ECF5 schema at the same time, based on the TylerEFM package
* Not adding to ECF4 because we're trying to move away from it, and any
time adding it there should be spent just making the move to ECF5.

Will make this PR, self review, and merge. Nothing to really fail or run
yet, won't be until some other ECF 5 code starts coming in, which it can
do soon